Update Rest test containers to be created in integration tests
